Fixed the gwt test cases - gwt-maven is not required on ./gwteventservice. This is just a java compiled classes and attached source in the distro (jar).
Also, stubbs have been added for maven scm ('mvn release' and snapshot 'mvn deploy') to a repository. We can use the svn repo for this purpose. This is commonly done in googlecode maven projects.

TODO - we don't have any of the sample webapps using this yet. They should really be on the build lifecycle too.